fix: apply namespace + caller edits missed from prior rename commit
Previous commit staged the git-mv rename but a bad pathspec in the same
git add invocation dropped the actual namespace/use-statement edits from
staging. This adds the content that was intended to ship with it.
