+++ ibl5/classes/Updater/Steps/CleanupPreseasonDataStep.php
@@ -8,6 +8,7 @@ use Boxscore\BoxscoreRepository;
 use League\LeagueContext;
 use Season\Season;
 use Updater\Contracts\PipelineStepInterface;
+use Updater\PreseasonCleanupRepository;
 use Updater\StepResult;
 
 /**
